UITEN-105 correctly show locales in native locale
Managing the translation values for locales was confusing because we
want to show "lang-A/lang-B" where lang-A is the locale in the current
locale and lang-B is the locale in that locale. For example, when the
current locale is English and the value is Russian, we want to show
`Russian / русский`, and when the value is Japanese, we want to show
`Japanese / 日本語`.

Rather than getting these values via translations, we can derive them
from react-intl directly.

Refs UITEN-105