No errors when inventory-related interfaces are not present
We now check for the presence of relevant interfaces at run-time, and
no make unsupported settings pages available when the relevant
interfaces are not present.

Those interfaces (all to do with inventory) are now marked as optional
in the package file.

It will be possible to simplify this code if/when support for the
`iface` element is added to the `<Settings>` component, but the
present fix does not depend on that future change: see
https://github.com/folio-org/stripes-smart-components/pull/1401#issuecomment-1771403292

Fixes UITEN-266.